improve test useability
