+++ .gitignore
@@ -162,4 +162,8 @@ cython_debug/
 .\#*
 *~
 
-examples/wandb/
+# Weight and Biases files
+*/wandb/
+
+# nohup output files
+*/nohup.out
