+++ JART_v1b_tests/w_max_and_w_min.yml
@@ -1,0 +1,49 @@
+Repeat_Times: 5
+USE_wandb: True
+CUDA_Enabled: False
+project_name: aihwkit_v1b_test
+epochs: 30
+batch_size: 64
+learning_rate:  0.05
+scheduler:
+  step_size: 10
+  gamma: 0.5
+
+pulse_related:
+  read_voltage: 0.2
+  pulse_voltage_SET: -0.342
+  pulse_voltage_RESET: 0.7065
+  pulse_length: 1.0e-6
+  base_time_step: 1.0e-8
+
+noise:
+  w_max:
+    device_to_device: 0.3
+  w_min:
+    device_to_device: 0.3
+  Ndiscmax:
+    device_to_device: 0.0
+    cycle_to_cycle_direct: 0.0
+    # 0 means disabled
+    upper_bound: 0.0
+    lower_bound: 0.0
+  Ndiscmin:
+    device_to_device: 0.0
+    cycle_to_cycle_direct: 0.0
+    # 0 means disabled
+    upper_bound: 0.0
+    lower_bound: 0.0
+  ldet:
+    device_to_device: 0.0
+    cycle_to_cycle_direct: 0.0
+    cycle_to_cycle_slope: 0.0
+    # 0 means disabled
+    upper_bound: 0.0
+    lower_bound: 0.0
+  rdet:
+    device_to_device: 0.0
+    cycle_to_cycle_direct: 0.0
+    cycle_to_cycle_slope: 0.0
+    # 0 means disabled
+    upper_bound: 0.0
+    lower_bound: 0.0
