--- examples/01_simple_layer_v1b.py
@@ -1,78 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# (C) Copyright 2020, 2021, 2022 IBM. All Rights Reserved.
-#
-# This code is licensed under the Apache License, Version 2.0. You may
-# obtain a copy of this license in the LICENSE.txt file in the root directory
-# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-#
-# Any modifications or derivative works of this code must retain this
-# copyright notice, and modified files need to carry a notice indicating
-# that they have been altered from the originals.
-
-"""aihwkit example 1: simple network with one layer.
-
-Simple network that consist of one analog layer. The network aims to learn
-to sum all the elements from one array.
-"""
-# pylint: disable=invalid-name
-
-# Imports from PyTorch.
-from torch import Tensor
-from torch.nn.functional import mse_loss
-
-# Imports from aihwkit.
-from aihwkit.nn import AnalogLinear
-from aihwkit.optim import AnalogSGD
-from aihwkit.simulator.configs import SingleRPUConfig
-from aihwkit.simulator.configs.devices import JARTv1bDevice
-from aihwkit.simulator.rpu_base import cuda
-
-import wandb
-
-learning_rate=0.1
-epochs = 100
-CUDA_Enabled = False
-
-
-if cuda.is_compiled() & CUDA_Enabled:
-    wandb.init(project="basic test v1b", group="Single Layer Perceptron", job_type="CUDA")
-else:
-    
-    wandb.init(project="basic test v1b", group="Single Layer Perceptron", job_type="CPU")
-
-wandb.config = {
-  "learning_rate": learning_rate,
-  "epochs": epochs,
-  "CUDA_Enabled": CUDA_Enabled
-}
-
-# Prepare the datasets (input and expected output).
-x = Tensor([[0.1, 0.2, 0.4, 0.3], [0.2, 0.1, 0.1, 0.3]])
-y = Tensor([[1.0, 0.5], [0.7, 0.3]])
-
-# Define a single-layer network, using a constant step device type.
-rpu_config = SingleRPUConfig(device=JARTv1bDevice())
-model = AnalogLinear(4, 2, bias=True,
-                     rpu_config=rpu_config)
-
-# Move the model and tensors to cuda if it is available.
-if cuda.is_compiled() & CUDA_Enabled:
-    x = x.cuda()
-    y = y.cuda()
-    model.cuda()
-
-# Define an analog-aware optimizer, preparing it for using the layers.
-opt = AnalogSGD(model.parameters(), lr=learning_rate)
-opt.regroup_param_groups(model)
-
-for epoch in range(epochs):
-    # Add the training Tensor to the model (input).
-    pred = model(x)
-    # Add the expected output Tensor.
-    loss = mse_loss(pred, y)
-    wandb.log({"loss": loss})
-    # Run training (backward propagation).
-    loss.backward()
-
-    opt.step()
